+++ app/src/main/java/com/ivoid/bj/CoinWallet.java
@@ -14,20 +14,20 @@ public class CoinWallet
 		editor = preference.edit();
 	}
 	
-	public void deposit(float amount) 
+	public void deposit(int amount)
 	{
-		editor.putFloat("CoinBalance", preference.getFloat("CoinBalance", 0f) + amount);
+		editor.putInt("CoinBalance", preference.getInt("CoinBalance", 0) + amount);
 		editor.commit();
 
 	}
 	
-	public void withdraw(float amount)
+	public void withdraw(int amount)
 	{
-		editor.putFloat("CoinBalance", preference.getFloat("CoinBalance", 0f) - amount);
+		editor.putInt("CoinBalance", preference.getInt("CoinBalance", 0) - amount);
 		editor.commit();
 	}
 	
 	//Accessors
-	public float getBalance()
-	{ return preference.getFloat("CoinBalance", 0f); }
+	public int getBalance()
+	{ return preference.getInt("CoinBalance", 0); }
 }
